make detect_lib return None instead of raising exception
